+++ package.json
@@ -15,6 +15,7 @@
     "node-fetch": "^1.6.3",
     "react": "^15.3.2",
     "react-addons-css-transition-group": "^15.4.2",
+    "react-autobind": "^1.0.6",
     "react-dom": "^15.3.2",
     "react-router": "4.0.0-alpha.3",
     "react-scripts": "^0.7.0"
